Added more error raising
Exception is a placeholder

+++ main.py
@@ -13,17 +13,17 @@ baseurl = 'https://api.twitch.tv/kraken/'
 
 class TwitchingWrapper():
     def __init__(self,token):
-        self.token = token
-        if self.token != None:
-            self.headers = {'Authorization':'OAuth ' + token}
-            dict1  = requests.get(baseurl + 'user', headers = self.headers)
-            dict1.raise_for_status()
-            self.dict1 = json.loads(dict1.text)
-            self.name = self.dict1 ['name']
-        else:
-            pass
+        if not isinstance(token, str) or not token:
+            raise Exception
+        self.headers = {'Authorization':'OAuth ' + token}
+        dict1  = requests.get(baseurl + 'user', headers = self.headers)
+        dict1.raise_for_status()
+        self.dict1 = json.loads(dict1.text)
+        self.name = self.dict1 ['name']
 
     def getblocklist(self,limit):
+        if not isinstance(limit, int) or not limit:
+            raise Exception
         params = {'limit':limit}
         time.sleep(waittime)
         dict1 = requests.get(baseurl + 'users/' + self.name + '/blocks', headers = self.headers,params = params)
@@ -31,12 +31,16 @@ class TwitchingWrapper():
         return json.loads(dict1.text)
         
     def putblocklist(self,target):
+        if not isinstance(target, str) or not target:
+            raise Exception
         time.sleep(waittime)
-        requests.put(baseurl + 'users/' + self.name + '/blocks/' + target,headers = self.headers).raise_for_status()
+        requests.put(baseurl + 'users/' + self.name + '/blocks/' + target, headers = self.headers).raise_for_status()
         
     def deleteblocklist(self,target):
+        if not isinstance(target, str) or not target:
+            raise Exception
         time.sleep(waittime)
-        requests.delete(baseurl + 'users/' + self.name + '/blocks/' + target,headers = self.headers).raise_for_status()
+        requests.delete(baseurl + 'users/' + self.name + '/blocks/' + target, headers = self.headers).raise_for_status()
                
     def getchannelinfo(self):
         time.sleep(waittime)
@@ -45,34 +49,52 @@ class TwitchingWrapper():
         return json.loads(dict1.text)
     
     def getchanneleditors(self,channel):
+        if not isinstance(channel, str) or not channel:
+            raise Exception
         time.sleep(waittime)
         dict1 = requests.get(baseurl + 'channels/' + channel +'/editors', headers = self.headers)
         dict1.raise_for_status()
         return json.loads(dict1.text)
     
     def getchannelfollowers(self,channel):
+        if not isinstance(channel, str) or not channel:
+            raise Exception
         time.sleep(waittime)
         dict1 = requests.get(baseurl + 'channels/' + channel +'/follows', headers = self.headers)
         dict1.raise_for_status()
         return json.loads(dict1.text)
     
     def getchannelvideos(self,channel):
+        if not isinstance(channel, str) or not channel:
+            raise Exception
         time.sleep(waittime)
         dict1 = requests.get(baseurl + 'channels/' + channel +'/videos', headers = self.headers)
         dict1.raise_for_status()
         return json.loads(dict1.text)
     
     def updatechannel(self,channel,status,game):
+        if not isinstance(channel, str) or not channel:
+            raise Exception
+        if not isinstance(status, str) or not status:
+            raise Exception
+        if not isinstance(game, str) or not game:
+            raise Exception
         params = {'status':status,'game':game}
         time.sleep(waittime)
         requests.put(baseurl + 'channels/' + channel, headers = self.headers, params = params).raise_for_status()
     
     def startcommercial(self,length,channel):
+        if not isinstance(length, int) or not length:
+            raise Exception
+        if not isinstance(channel, str) or not channel:
+            raise Exception
         parms = {'channel_commercial':length}
         time.sleep(waittime)
         requests.post(baseurl + 'channels/' + channel + '/commercial', headers = self.headers, parms = parms).raise_for_status()
     
     def resetstreamkey(self,channel):
+        if not isinstance(channel, str) or not channel:
+            raise Exception
         time.sleep(waittime)
         requests.delete(baseurl + 'channels/' + channel + '/stream_key', headers = self.headers).raise_for_status()
